requirements.txt
pandas
openpyxl
numpy
faiss-cpu
tiktoken
python-dotenv
openai